requirements.txt
aiomysql==0.2.0
python-dotenv==1.0.0
pyyaml==6.0